Updated dependencies. Refactored a lot of code, making more use of exceptions, especially in Creep methods. Modularized some async actions. Changed hauling into an implementation making more use of async and exceptions in a clean way. Added more DEBUG constants used to conditionally turn on some debug strings with no runtime cost when off. Updated the code to compile with the latest rustc nightly.

+++ src/towers.rs
@@ -2,6 +2,6 @@ use std::cmp::{max, min};
 use screeps::{TOWER_FALLOFF, TOWER_FALLOFF_RANGE, TOWER_OPTIMAL_RANGE, TOWER_POWER_ATTACK};
 
 pub fn tower_attack_power(dist: u8) -> u16 {
-    let effective_dist = max(TOWER_OPTIMAL_RANGE, min(TOWER_FALLOFF_RANGE, dist as u32));
-    (TOWER_POWER_ATTACK - ((TOWER_POWER_ATTACK as f32 * TOWER_FALLOFF) as u32) * (effective_dist - TOWER_OPTIMAL_RANGE) / (TOWER_FALLOFF_RANGE - TOWER_OPTIMAL_RANGE)) as u16
+    let effective_dist = max(TOWER_OPTIMAL_RANGE, min(TOWER_FALLOFF_RANGE, dist));
+    (TOWER_POWER_ATTACK - ((TOWER_POWER_ATTACK as f64 * TOWER_FALLOFF) as u32) * ((effective_dist - TOWER_OPTIMAL_RANGE) / (TOWER_FALLOFF_RANGE - TOWER_OPTIMAL_RANGE)) as u32) as u16
 }
